Process OCR XML using ElementTree
This experiment saves about 20% over the SAX interface. 
Unfortunately, it adds a bit more assumption about the
structure of the document which might be a problem if there
is some malformed or nested ALTO content.

lxml was tested but the faster query times weren't balanced
by the longer parse times from libxml2. Unfortunately, XPath
isn't really a time-savings because we need to retrieve
multiple attributes and that's unsafe to do with the libxml2
XPath 1.0 API.